fixes for new columns not having reactions
