+++ app/src/state/initialize.ts
@@ -11,18 +11,18 @@ export async function initialize() {
   actions.fetchFaciesPatterns();
   actions.fetchTimescaleDataAction();
   actions.fetchAllWorkshops();
-  const datapack = (await actions.fetchDatapack({
-    title: "TimeScale Creator Internal Datapack",
-    type: "official",
-    isPublic: true
-  }))!;
-  actions.addDatapack(datapack);
-  actions.processDatapackConfig([
-    {
-      title: datapack.title,
-      isPublic: datapack.isPublic,
-      storedFileName: datapack.storedFileName,
-      ...extractDatapackType(datapack)
-    }
-  ]);
+  // const datapack = (await actions.fetchDatapack({
+  //   title: "TimeScale Creator Internal Datapack",
+  //   type: "official",
+  //   isPublic: true
+  // }))!;
+  // actions.addDatapack(datapack);
+  // actions.processDatapackConfig([
+  //   {
+  //     title: datapack.title,
+  //     isPublic: datapack.isPublic,
+  //     storedFileName: datapack.storedFileName,
+  //     ...extractDatapackType(datapack)
+  //   }
+  // ]);
 }
