+++ app/src/util/util.ts
@@ -229,18 +229,6 @@ export function findSerialNum(name: string) {
   return serialNumber;
 }
 
-export function convertColumnInfoToRenderColumnInfo(column: ColumnInfo): RenderColumnInfo {
-  const renderColumn: RenderColumnInfo = observable(
-    {
-      ...column,
-      columnRef: column,
-      children: column.children.map((child) => child.name)
-    },
-    { columnRef: false }
-  );
-  return renderColumn;
-}
-
 export function getChildRenderColumns(
   parent: RenderColumnInfo,
   columnHashMap: Map<string, RenderColumnInfo>
